+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## master
 
+#### News
+
 - Prints pipe info into a table on the console;
+- The previous result is keeped alive for further checks;
 
 ## v0.6.0